Refactor payment method modules
