+++ Sources/PrimerSDK/Classes/User Interface/Countries/CountrySelectorViewController.swift
@@ -2,32 +2,60 @@
 
 import UIKit
 
-internal protocol SearchableItemsPaymentMethodTokenizationViewModelProtocol {
-    
-    var tableView: UITableView { get set }
-    var searchableTextField: PrimerSearchTextField { get set }
-    
-    func cancel()
-}
-
 internal class CountrySelectorViewController: PrimerFormViewController {
     
     let theme: PrimerThemeProtocol = DependencyContainer.resolve()
     
-    private var delegate: SearchableItemsPaymentMethodTokenizationViewModelProtocol!
     private let countries = CountryCode.allCases
+    var dataSource = CountryCode.allCases {
+        didSet {
+            tableView.reloadData()
+        }
+    }
+    
     internal private(set) var subtitle: String?
-    private var paymentMethod: PrimerPaymentMethod
+    private var paymentMethodType: String
+    
+    internal var didSelectCountryCode: ((_ countryCode: CountryCode) -> Void)?
+    
+    internal lazy var tableView: UITableView = {
+        let theme: PrimerThemeProtocol = DependencyContainer.resolve()
+        
+        let tableView = UITableView()
+        tableView.showsVerticalScrollIndicator = false
+        tableView.showsHorizontalScrollIndicator = false
+        tableView.backgroundColor = theme.view.backgroundColor
+        
+        if #available(iOS 11.0, *) {
+            tableView.contentInsetAdjustmentBehavior = .never
+        }
+
+        tableView.rowHeight = 41
+        tableView.register(CountryTableViewCell.self, forCellReuseIdentifier: CountryTableViewCell.className)
+        tableView.dataSource = self
+        tableView.delegate = self
+        return tableView
+    }()
+    
+    internal lazy var searchableTextField: PrimerSearchTextField = {
+        let textField = PrimerSearchTextField(frame: .zero)
+        textField.translatesAutoresizingMaskIntoConstraints = false
+        textField.heightAnchor.constraint(equalToConstant: 35).isActive = true
+        textField.delegate = self
+        textField.borderStyle = .none
+        textField.layer.cornerRadius = 3.0
+        textField.font = UIFont.systemFont(ofSize: 16.0)
+        textField.placeholder = Strings.CountrySelector.searchCountryTitle
+        textField.rightViewMode = .always
+        return textField
+    }()
     
     deinit {
-        delegate.cancel()
-        delegate = nil
         log(logLevel: .debug, message: "🧨 deinit: \(self) \(Unmanaged.passUnretained(self).toOpaque())")
     }
     
-    init(delegate: SearchableItemsPaymentMethodTokenizationViewModelProtocol, paymentMethod: PrimerPaymentMethod) {
-        self.delegate = delegate
-        self.paymentMethod = paymentMethod
+    init(paymentMethodType: String) {
+        self.paymentMethodType = paymentMethodType
         super.init(nibName: nil, bundle: nil)
     }
     
@@ -38,13 +66,15 @@ internal class CountrySelectorViewController: PrimerFormViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        self.verticalStackView.addArrangedSubview(self.tableView)
+        
         let viewEvent = Analytics.Event(
             eventType: .ui,
             properties: UIEventProperties(
                 action: .view,
                 context: Analytics.Event.Property.Context(
                     issuerId: nil,
-                    paymentMethodType: self.paymentMethod.type,
+                    paymentMethodType: paymentMethodType,
                     url: nil),
                 extra: nil,
                 objectType: .view,
@@ -55,8 +85,8 @@ internal class CountrySelectorViewController: PrimerFormViewController {
 
         view.backgroundColor = theme.view.backgroundColor
         view.translatesAutoresizingMaskIntoConstraints = false
-        view.heightAnchor.constraint(equalToConstant: 120+(CGFloat(countries.count)*delegate.tableView.rowHeight)).isActive = true
-        delegate.tableView.isScrollEnabled = false
+        view.heightAnchor.constraint(equalToConstant: 120+(CGFloat(countries.count)*self.tableView.rowHeight)).isActive = true
+        self.tableView.isScrollEnabled = false
                 
         verticalStackView.spacing = 5
         
@@ -68,15 +98,74 @@ internal class CountrySelectorViewController: PrimerFormViewController {
         bankTitleLabel.textColor = theme.text.title.color
         verticalStackView.addArrangedSubview(bankTitleLabel)
         
-        verticalStackView.addArrangedSubview(delegate.searchableTextField)
+        verticalStackView.addArrangedSubview(self.searchableTextField)
                 
         let separator = UIView()
         separator.translatesAutoresizingMaskIntoConstraints = false
         separator.heightAnchor.constraint(equalToConstant: 5).isActive = true
         verticalStackView.addArrangedSubview(separator)
         
-        self.verticalStackView.addArrangedSubview(self.delegate.tableView)
-        self.delegate.tableView.translatesAutoresizingMaskIntoConstraints = false
+        self.tableView.translatesAutoresizingMaskIntoConstraints = false
+    }
+}
+
+extension CountrySelectorViewController: UITableViewDataSource, UITableViewDelegate {
+    
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return dataSource.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let country = dataSource[indexPath.row]
+        let cell = tableView.dequeueReusableCell(withIdentifier: CountryTableViewCell.className, for: indexPath) as! CountryTableViewCell
+        cell.configure(viewModel: country)
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let countryCode = self.dataSource[indexPath.row]
+        self.didSelectCountryCode?(countryCode)
+    }
+}
+
+extension CountrySelectorViewController: UITextFieldDelegate {
+    
+    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        if string == "\n" {
+            // Keyboard's return button tapoped
+            textField.resignFirstResponder()
+            return false
+        }
+        
+        var query: String
+        
+        if string.isEmpty {
+            query = String((textField.text ?? "").dropLast())
+        } else {
+            query = (textField.text ?? "") + string
+        }
+        
+        if query.isEmpty {
+            dataSource = countries
+            return true
+        }
+        
+        var countryResults: [CountryCode] = []
+        
+        for country in countries {
+            if country.country.lowercased().folding(options: .diacriticInsensitive, locale: nil).contains(query.lowercased().folding(options: .diacriticInsensitive, locale: nil)) == true {
+                countryResults.append(country)
+            }
+        }
+        
+        dataSource = countryResults
+        
+        return true
+    }
+    
+    func textFieldShouldClear(_ textField: UITextField) -> Bool {
+        dataSource = countries
+        return true
     }
 }
