+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/Tokenization Modules/ApayaTokenizationModule.swift
@@ -60,10 +60,10 @@ class ApayaTokenizationModule: TokenizationModule {
         }
     }
     
-    override func start() -> Promise<PrimerPaymentMethodTokenData> {
+    override func startFlow() -> Promise<PrimerPaymentMethodTokenData> {
 //        NotificationCenter.default.addObserver(self, selector: #selector(self.receivedNotification(_:)), name: Notification.Name.urlSchemeRedirect, object: nil)
         
-        return super.start()
+        return super.startFlow()
     }
     
     override func performPreTokenizationSteps() -> Promise<Void> {
