+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/Tokenization Modules/ApplePayTokenizationModule.swift
@@ -68,10 +68,10 @@ class ApplePayTokenizationModule: TokenizationModule {
         }
     }
     
-    override func start() -> Promise<PrimerPaymentMethodTokenData> {
+    override func startFlow() -> Promise<PrimerPaymentMethodTokenData> {
 //        NotificationCenter.default.addObserver(self, selector: #selector(self.receivedNotification(_:)), name: Notification.Name.urlSchemeRedirect, object: nil)
         
-        return super.start()
+        return super.startFlow()
     }
     
     override func performPreTokenizationSteps() -> Promise<Void> {
