+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/Tokenization Modules/BankSelectorTokenizationModule.swift
@@ -11,46 +11,9 @@ import Foundation
 
 class BankSelectorTokenizationModule: TokenizationModule {
     
-    var banks: [AdyenBank] = []
-    private var dataSource: [AdyenBank] = [] {
-        didSet {
-            tableView.reloadData()
-        }
-    }
+    private var banks: [AdyenBank] = []
     private var bankSelectionCompletion: ((AdyenBank) -> Void)?
     private var selectedBank: AdyenBank?
-    
-    internal lazy var tableView: UITableView = {
-        let theme: PrimerThemeProtocol = DependencyContainer.resolve()
-        
-        let tableView = UITableView()
-        tableView.showsVerticalScrollIndicator = false
-        tableView.showsHorizontalScrollIndicator = false
-        tableView.backgroundColor = theme.view.backgroundColor
-        
-        if #available(iOS 11.0, *) {
-            tableView.contentInsetAdjustmentBehavior = .never
-        }
-
-        tableView.rowHeight = 41
-        tableView.register(BankTableViewCell.self, forCellReuseIdentifier: BankTableViewCell.identifier)
-        tableView.dataSource = self
-        tableView.delegate = self
-        return tableView
-    }()
-    
-    internal lazy var searchBankTextField: PrimerSearchTextField? = {
-        let textField = PrimerSearchTextField(frame: .zero)
-        textField.translatesAutoresizingMaskIntoConstraints = false
-        textField.heightAnchor.constraint(equalToConstant: 35).isActive = true
-        textField.delegate = self
-        textField.borderStyle = .none
-        textField.layer.cornerRadius = 3.0
-        textField.font = UIFont.systemFont(ofSize: 16.0)
-        textField.placeholder = Strings.BankSelector.searchBankTitle
-        textField.rightViewMode = .always
-        return textField
-    }()
     
     override func validate() -> Promise<Void> {
         return Promise { seal in
@@ -64,10 +27,10 @@ class BankSelectorTokenizationModule: TokenizationModule {
         }
     }
     
-    override func start() -> Promise<PrimerPaymentMethodTokenData> {
+    override func startFlow() -> Promise<PrimerPaymentMethodTokenData> {
 //        NotificationCenter.default.addObserver(self, selector: #selector(self.receivedNotification(_:)), name: Notification.Name.urlSchemeRedirect, object: nil)
         
-        return super.start()
+        return super.startFlow()
     }
     
     override func performPreTokenizationSteps() -> Promise<Void> {
@@ -102,7 +65,6 @@ class BankSelectorTokenizationModule: TokenizationModule {
             }
             .then { banks -> Promise<Void> in
                 self.banks = banks
-                self.dataSource = banks
                 return self.presentPaymentMethodUserInterface()
             }
             .then { () -> Promise<Void> in
@@ -112,6 +74,11 @@ class BankSelectorTokenizationModule: TokenizationModule {
                 return self.awaitUserInput()
             }
             .then { () -> Promise<Void> in
+                DispatchQueue.main.async {
+                    PrimerUIManager.primerRootViewController?.showLoadingScreenIfNeeded(
+                        imageView: self.paymentMethodModule.userInterfaceModule.makeIconImageView(withDimension: 24.0),
+                        message: nil)
+                }
                 return self.firePrimerWillCreatePaymentEvent(PrimerPaymentMethodData(type: self.paymentMethodModule.paymentMethodConfiguration.type))
             }
             .done {
@@ -138,14 +105,6 @@ class BankSelectorTokenizationModule: TokenizationModule {
             .done {
                 seal.fulfill()
             }
-//            .ensure { [unowned self] in
-//                DispatchQueue.main.async {
-////                    self.willDismissPaymentMethodUI?()
-//                    self.webViewController?.dismiss(animated: true, completion: {
-////                        self.didDismissPaymentMethodUI?()
-//                    })
-//                }
-//            }
             .catch { err in
                 seal.reject(err)
             }
@@ -229,11 +188,15 @@ class BankSelectorTokenizationModule: TokenizationModule {
     private func presentPaymentMethodUserInterface() -> Promise<Void> {
         return Promise { seal in
             DispatchQueue.main.async {
-                let bsvc = BankSelectorViewController(tokenizationModule: self)
-                DispatchQueue.main.async {
-                    PrimerUIManager.primerRootViewController?.show(viewController: bsvc)
-                    seal.fulfill()
+                let bsvc = self.paymentMethodModule.userInterfaceModule.createBanksSelectorViewController(with: self.banks)
+                
+                bsvc.didSelectBank = { [weak self] bank in
+                    self?.selectedBank = bank
+                    self?.bankSelectionCompletion?(bank)
                 }
+                
+                PrimerUIManager.primerRootViewController?.show(viewController: bsvc)
+                seal.fulfill()
             }
         }
     }
@@ -241,70 +204,9 @@ class BankSelectorTokenizationModule: TokenizationModule {
     private func awaitUserInput() -> Promise<Void> {
         return Promise { seal in
             self.bankSelectionCompletion = { bank in
-                self.selectedBank = bank
                 seal.fulfill()
             }
         }
-    }
-}
-
-extension BankSelectorTokenizationModule: UITableViewDataSource, UITableViewDelegate {
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return dataSource.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let bank = dataSource[indexPath.row]
-        let cell = tableView.dequeueReusableCell(withIdentifier: "BankTableViewCell", for: indexPath) as! BankTableViewCell
-        cell.configure(viewModel: bank)
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let bank = self.dataSource[indexPath.row]
-        self.bankSelectionCompletion?(bank)
-    }
-}
-
-extension BankSelectorTokenizationModule: UITextFieldDelegate {
-    
-    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
-        if string == "\n" {
-            // Keyboard's return button tapoped
-            textField.resignFirstResponder()
-            return false
-        }
-        
-        var query: String
-        
-        if string.isEmpty {
-            query = String((textField.text ?? "").dropLast())
-        } else {
-            query = (textField.text ?? "") + string
-        }
-        
-        if query.isEmpty {
-            dataSource = banks
-            return true
-        }
-        
-        var bankResults: [AdyenBank] = []
-        
-        for bank in banks {
-            if bank.name.lowercased().folding(options: .diacriticInsensitive, locale: nil).contains(query.lowercased().folding(options: .diacriticInsensitive, locale: nil)) == true {
-                bankResults.append(bank)
-            }
-        }
-        
-        dataSource = bankResults
-        
-        return true
-    }
-    
-    func textFieldShouldClear(_ textField: UITextField) -> Bool {
-        dataSource = banks
-        return true
     }
 }
