+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/Tokenization Modules/PrimerTestPaymentMethodTokenizationModule.swift
@@ -53,10 +53,10 @@ class PrimerTestPaymentMethodTokenizationModule: TokenizationModule {
         }
     }
     
-    override func start() -> Promise<PrimerPaymentMethodTokenData> {
+    override func startFlow() -> Promise<PrimerPaymentMethodTokenData> {
 //        NotificationCenter.default.addObserver(self, selector: #selector(self.receivedNotification(_:)), name: Notification.Name.urlSchemeRedirect, object: nil)
         
-        return super.start()
+        return super.startFlow()
     }
     
     override func performPreTokenizationSteps() -> Promise<Void> {
