+++ Sources/PrimerSDK/Classes/User Interface/TokenizationViewModels/Tokenization Modules/VaultedPaymentMethodTokenizationModule.swift
@@ -34,10 +34,10 @@ class VaultedPaymentMethodTokenizationModule: TokenizationModule {
         }
     }
     
-    override func start() -> Promise<PrimerPaymentMethodTokenData> {
+    override func startFlow() -> Promise<PrimerPaymentMethodTokenData> {
 //        NotificationCenter.default.addObserver(self, selector: #selector(self.receivedNotification(_:)), name: Notification.Name.urlSchemeRedirect, object: nil)
         
-        return super.start()
+        return super.startFlow()
     }
     
     override func performPreTokenizationSteps() -> Promise<Void> {
